TIP-874: move FileStorage constant
