+++ src/Akeneo/Tool/Bundle/ApiBundle/tests/integration/Controller/Media/DownloadMediaFileIntegration.php
@@ -4,7 +4,7 @@ namespace Akeneo\Tool\Bundle\ApiBundle\tests\integration\Controller\MediaFile;
 
 use Akeneo\Tool\Component\FileStorage\Model\FileInfo;
 use Akeneo\Tool\Bundle\ApiBundle\tests\integration\Controller\Media\AbstractMediaFileTestCase;
-use Pim\Component\Catalog\FileStorage;
+use Akeneo\Pim\Enrichment\Component\FileStorage;
 use Symfony\Component\HttpFoundation\Response;
 
 class DownloadMediaFileIntegration extends AbstractMediaFileTestCase
