+++ src/Pim/Bundle/EnrichBundle/spec/Imagine/Loader/FlysystemLoaderSpec.php
@@ -5,7 +5,7 @@ namespace spec\Pim\Bundle\EnrichBundle\Imagine\Loader;
 use Akeneo\Tool\Component\FileStorage\FilesystemProvider;
 use League\Flysystem\FilesystemInterface;
 use PhpSpec\ObjectBehavior;
-use Pim\Component\Catalog\FileStorage;
+use Akeneo\Pim\Enrichment\Component\FileStorage;
 
 class FlysystemLoaderSpec extends ObjectBehavior
 {
